fix v9 parsing by preserving the complete packets during buffering
aggregating the data flowsets does not work, because all records are based on the packet's timestamp
to simplify parsing in the codec, the aggregator now collects all templates/option templates into a protobuf
and adds received and buffered data flows to be parsed.
the netflow packets are preserved completely and can also contain templates. the codec will not use them, though, only the aggregator does