Improved NetFlowV9 support (#21)
* move template caching and flow buffering in custom message aggregator

much of this is still wip

* remove template cache

* wip for codec aggregation and parsing custom format of v9

currently fixing tests

* wip

* migrate tests

* fix v9 parsing by preserving the complete packets during buffering

aggregating the data flowsets does not work, because all records are based on the packet's timestamp
to simplify parsing in the codec, the aggregator now collects all templates/option templates into a protobuf
and adds received and buffered data flows to be parsed.
the netflow packets are preserved completely and can also contain templates. the codec will not use them, though, only the aggregator does

* remove duplicated license header

* update protobuf comment

* update comment

* tweak license header to avoid diff

* fix guice setup for transport

* fix handler setup

the codec-aggregator is null in super class, so that the put call added it after the raw-message handler
thus the code never ran and screwed up parsing

* change how the packet cache works

the previous implementation only checked for a single template id to be present for each packet, which in general is wrong if not all templates arrive at the same time (which might happen for large numbers of active templates)
the new implementation manually checks each packet's template requirements agains the ids of received templates, for the current remoteaddress/source id combination

* prefix netflow v9 fields with nf_

* remove unused optional template atomic ref

* don't prefix unknown fields names, that is done centrally now

fixes double prefixing of fields

* don't forget to fix test

* add flow timestamp fields from ipfix

* fix test after field definition list update

(cherry picked from commit 4e9d68fdf435de0b55a6852078ef557611bcaaef)